Use scrypt for password hashing. Require 7 emoji.
Changing the emoji minimum:

With 1,791 emoji, each emoji character provides approximately 10.8-bits of
entropy. If a password cracker can make 1,000,000,000,000 guesses per second
(only need 5 systems of 8 Nvidia GTX 1080 GUPs each running Hashcat 3.0), then
for the attacker to be at least 50% sure he found the correct sequence, we can
expect it to take:

    55-bits: ~ 5 hours.
    60-bits: ~ 7 days,
    65-bits: ~ 7 months.
    70-bits: ~ 19 years.
    75-bits: ~ 600 years.
    80-bits: ~ 19 millenia.

As such, the default of 4 emoji only provides about 43-bits of entropy. Knowing
the statistics above, means we need to increase our search space. 6 emoji
provide about 65-bits of entropy, and 7 emoji provide about 76-bits of entropy.

Changing the password hashing function:

Passwords should never be hashed with anything other than a dedicated password
hashing function. This means not using generic cryptographic hashing functions
like SHA-256 or SHA-512. Instead, the password should be hashed, on order of
preference, with:

    * scrypt
    * bcrypt
    * sha512crypt
    * sha256crypt
    * PBKDF2
    * Argon2 (recent PHC winner- too new)

This push changes the default of:

    base64(sha256(sha512(password)))

With:

    base64(scrypt(password))

In this implementation, it uses the recomended settings by Colin Percival of
N=14, r=8, p=1, to provide a 16MB memory cost in addition to the CPU cost. In
my testing on my Intel Core 2 Duo, it takes about 5ms to completion.